Prototype 2 - In Progress
Created prototype for the home page (landing page) and partially the login page for students.
A problem encountered while designing was that Figma does not allow the same component to have two events (hover and on click). For this purpose,  two buttons are added on the home page just for design purpose. But the actual website will have one button with both interactions.